# API-SPECIFICATION.yml
openapi: 3.0.0
paths:
...
  /api/v1/restaurants/{id}/menu:
    get:
      operationId: RestaurantController_getRestaurantMenu
      summary: Get restaurant menu
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
        -
          name: category
          required: false
          in: query
          schema:
            type: string
      responses:
        200:
          description: Menu retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurants (Customer)
      security:
        -
          JWT-auth: []
  /api/v1/restaurant/profile:
    get:
      operationId: RestaurantOwnerController_getProfile
      summary: Get own restaurant profile
      parameters: []
      responses:
        200:
          description: Restaurant profile retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
    post:
      operationId: RestaurantOwnerController_createProfile
      summary: Create restaurant profile (first-time setup)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantDto"
      responses:
        201:
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        401:
          description: Unauthorized
        409:
          description: Restaurant already exists
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
    patch:
      operationId: RestaurantOwnerController_updateProfile
      summary: Update restaurant profile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantDto"
      responses:
        200:
          description: Restaurant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
  /api/v1/restaurant/status:
    patch:
      operationId: RestaurantOwnerController_updateStatus
      summary: Toggle restaurant open/closed status
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantStatusDto"
      responses:
        200:
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        401:
          description: Unauthorized
        403:
          description: Restaurant not approved or not active
        404:
          description: Restaurant not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
  /api/v1/restaurant/menu:
    get:
      operationId: RestaurantOwnerController_getMenu
      summary: Get own restaurant menu
      parameters: []
      responses:
        200:
          description: Menu retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
    post:
      operationId: RestaurantOwnerController_createMenuItem
      summary: Create menu item
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemDto"
      responses:
        201:
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        401:
          description: Unauthorized
        404:
          description: Restaurant not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
  /api/v1/restaurant/menu/{id}:
    patch:
      operationId: RestaurantOwnerController_updateMenuItem
      summary: Update menu item
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuItemDto"
      responses:
        200:
          description: Menu item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        401:
          description: Unauthorized
        404:
          description: Menu item not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
    delete:
      operationId: RestaurantOwnerController_deleteMenuItem
      summary: Delete menu item
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        204:
          description: Menu item deleted
        401:
          description: Unauthorized
        404:
          description: Menu item not found
      tags:
        - Restaurant Management (Owner)
      security:
        -
          JWT-auth: []
  /api/v1/restaurant/orders:
    get:
      operationId: RestaurantOrderController_getRestaurantOrders
      summary: Get orders for restaurant
      parameters: []
      responses:
        200:
          description: Orders retrieved successfully
      tags:
        - Restaurant Orders
      security:
        -
          JWT-auth: []
  /api/v1/restaurant/orders/{id}/accept:
    patch:
      operationId: RestaurantOrderController_acceptOrder
      summary: Accept order and trigger driver assignment
      parameters:
        -
          name: id
          required: true
          in: path
          description: Order ID
          schema:
            type: string
      responses:
        200:
          description: Order accepted and driver assigned successfully
        400:
          description: Invalid order status
        403:
          description: Order does not belong to this restaurant
        404:
          description: Order not found
        503:
          description: No drivers available
      tags:
        - Restaurant Orders
      security:
        -
          JWT-auth: []
...
  /api/v1/restaurant/orders/{id}/reject:
    patch:
      operationId: RestaurantOrderController_rejectOrder
      summary: Reject order
      parameters:
        -
          name: id
          required: true
          in: path
          description: Order ID
          schema:
            type: string
      responses:
        200:
          description: Order rejected successfully
        400:
          description: Invalid order status
        403:
          description: Order does not belong to this restaurant
        404:
          description: Order not found
      tags:
        - Restaurant Orders
      security:
        -
          JWT-auth: []
...
info:
  title: SwiftEats API
  description: Real-Time Food Delivery Platform API
  version: 1.0
  contact:
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    Restaurant:
      type: object
      properties:
    MenuItem:
      type: object
      properties:
    CreateRestaurantDto:
      type: object
      properties:
        name:
          type: string
          example: The Great Indian Kitchen
          description: Restaurant name
        description:
          type: string
          example: Authentic Indian cuisine with a modern twist
          description: Restaurant description
        phone:
          type: string
          example: +919876543210
          description: Contact phone number
        email:
          type: string
          example: restaurant@example.com
          description: Contact email
        address:
          type: string
          example: 123 MG Road, Brigade Road
          description: Full address
        city:
          type: string
          example: Mumbai
          description: City name
        state:
          type: string
          example: Maharashtra
          description: State name
        postalCode:
          type: string
          example: 400001
          description: Postal code
        latitude:
          type: number
          example: 19.076
          description: Latitude (-90 to 90)
        longitude:
          type: number
          example: 72.8777
          description: Longitude (-180 to 180)
        cuisineType:
          type: string
          example: Indian
          description: Type of cuisine
        openingTime:
          type: string
          example: "09:00"
          description: "Opening time (HH:MM format)"
        closingTime:
          type: string
          example: "22:00"
          description: "Closing time (HH:MM format)"
        averagePreparationTimeMins:
          type: number
          example: 30
          description: Average preparation time in minutes
        imageUrl:
          type: string
          example: "https://example.com/restaurant-image.jpg"
          description: Restaurant image URL
      required:
        - name
        - phone
        - address
        - city
        - latitude
        - longitude
        - openingTime
        - closingTime
    UpdateRestaurantDto:
      type: object
      properties:
        name:
          type: string
          example: The Great Indian Kitchen
          description: Restaurant name
        description:
          type: string
          example: Authentic Indian cuisine with a modern twist
          description: Restaurant description
        phone:
          type: string
          example: +919876543210
          description: Contact phone number
        email:
          type: string
          example: restaurant@example.com
          description: Contact email
        address:
          type: string
          example: 123 MG Road, Brigade Road
          description: Full address
        city:
          type: string
          example: Mumbai
          description: City name
        state:
          type: string
          example: Maharashtra
          description: State name
        postalCode:
          type: string
          example: 400001
          description: Postal code
        latitude:
          type: number
          example: 19.076
          description: Latitude (-90 to 90)
        longitude:
          type: number
          example: 72.8777
          description: Longitude (-180 to 180)
        cuisineType:
          type: string
          example: Indian
          description: Type of cuisine
        openingTime:
          type: string
          example: "09:00"
          description: "Opening time (HH:MM format)"
        closingTime:
          type: string
          example: "22:00"
          description: "Closing time (HH:MM format)"
        averagePreparationTimeMins:
          type: number
          example: 30
          description: Average preparation time in minutes
        imageUrl:
          type: string
          example: "https://example.com/restaurant-image.jpg"
          description: Restaurant image URL
    UpdateRestaurantStatusDto:
      type: object
      properties:
        isOpen:
          type: boolean
          example: true
          description: Open/Close restaurant
      required:
        - isOpen
    CreateMenuItemDto:
      type: object
      properties:
        name:
          type: string
          example: Butter Chicken
          description: Menu item name
        description:
          type: string
          example: Tender chicken in rich tomato and butter sauce
          description: Item description
        price:
          type: number
          example: 350
          description: Price in INR
        category:
          type: string
          example: Main Course
          description: Category of the item
        isAvailable:
          type: boolean
          example: true
          description: Is item available
        isVegetarian:
          type: boolean
          example: true
          description: Is item vegetarian
        isVegan:
          type: boolean
          example: false
          description: Is item vegan
        imageUrl:
          type: string
          example: "https://example.com/butter-chicken.jpg"
          description: Item image URL
        preparationTimeMins:
          type: number
          example: 20
          description: Preparation time in minutes
      required:
        - name
        - price
        - category
    UpdateMenuItemDto:
      type: object
      properties:
        name:
          type: string
          example: Butter Chicken
          description: Menu item name
        description:
          type: string
          example: Tender chicken in rich tomato and butter sauce
          description: Item description
        price:
          type: number
          example: 350
          description: Price in INR
        category:
          type: string
          example: Main Course
          description: Category of the item
        isAvailable:
          type: boolean
          example: true
          description: Is item available
        isVegetarian:
          type: boolean
          example: true
          description: Is item vegetarian
        isVegan:
          type: boolean
          example: false
          description: Is item vegan
        imageUrl:
          type: string
          example: "https://example.com/butter-chicken.jpg"
          description: Item image URL
        preparationTimeMins:
          type: number
          example: 20
          description: Preparation time in minutes
...